tests/wpt/tests/navigation-api/precommit-handler/precommitHandler-window-stop-before-commit.html

<!doctype html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../navigation-methods/return-value/resources/helpers.js"></script>
<body>
<script type="module">
import { ensureWindowLoadEventFired } from "../resources/helpers.mjs";

promise_test(async t => {
  // Wait for after the load event because window.stop() hangs the test harness
  // if called before the load event.
  await ensureWindowLoadEventFired(t);

  navigation.onnavigate = e => {
    e.intercept({ precommitHandler: async () => {} });
  };

  navigation.onnavigatesuccess = t.unreached_func("navigatesuccess must not fire");
  let navigateerror_called = false;
  navigation.onnavigateerror = t.step_func(() => {
    navigateerror_called = true;
    assert_equals(location.hash, "");
  });

  let promises = navigation.navigate("#ShouldNotCommit");
  assert_equals(location.hash, "");

  window.stop();
  await assertBothRejectDOM(t, promises, "AbortError");

  assert_equals(location.hash, "");
  assert_true(navigateerror_called);
}, " precommitHandler with window.stop() before commit");
</script>
</body>
